Introduce courses resource and controller
This is the client-facing side of the service. As per the readme, we
support a courses.json route with an optional query string.

Tests exercise the endpoint options that are outlined in the readme. The
CourseRepository that it's interacting with is nothing more than a stub
right now. Building that is next.

Haven't implemented views yet. So if you try to hit this endpoint with a
browser you'll get errors.