Refactor courses_controller to extract search obj
There was clearly another object in the controller action. The checking
for query strings, the parsing of the query string, the seeing if
CourseRepo will handle the query, etc. None of that seemed like
knowledge I wanted in the controller.

So, I've removed that logic (which will be re-implemented later) and
introduced a simple collaborator for the controller to work with
course_search.