Premptively change api of QueryParse.parse
Thinking ahead, the point of the `valid?` method is to say if the search
target can actually respond to the search. If it can, run the search, if
not, return nil.

This means we need to pass the search target (the course_repository, in
this case) into the query parser.