Implement actual querying for le courses
This class does the actual querying for courses and also handles the
caching. This wasn't the initial plan. My thought was that I could pass
in different backends to the CourseRepository, one for an actual DB
backend and one for a in-memory backend.

But, while implementing this I found I could do both in this single file
and didn't see a big reason to split the two. Maybe that will become
necessary later.

The class-level finders call to retrieve with a sql fragment. The
retrieve method checks the cache for a key with that fragment. If yes,
it returns that. If not, it runs the full query and then caches the
results.

Using ActiveRecords find_by_sql here helps a lot because the results are
returned with getters like 'course_id', etc. Doing a raw
ActiveRecord::Base.connection.execute(sql) doesn't do that. It just
returns a collection of arrays which then have to be parsed.

This is not currently tested. Since all it does is make sql queries,
to external dbs, it's not a great candidate for unit testing. But the
integration tests should ensure that it's working.