Add end-to-end request spec
This test file, while slow, does the full end-to-end test starting from
a client hitting a URL all the way to us returning JSON that we have
retrieved from the database. It's the one test that actually exercises the
db code that's in app/models/liberal_education_course.rb.

But, yes, it takes ~14 seconds to run (at least while I'm remote, may be
faster in the office). This is because it can't take advantage of the
caching that's part of liberal_education_course. It's not something you want to run all the
time. We need to investigate a way to run this as part of deployment or
PR, but not run it as part of the default rake task.

As part of the test, I've updated the readme to reflect what actually
happens.