Convert to using an application db account
I had been testing this using the asr_warehouse db account, but that's
not a long-term solution. We now have the asr_le_service account which
is granted access to the tables it needs

https://github.umn.edu/asrweb/asr_warehouse/commit/94909d66e020b8442f0ae1bca7305d00b7f348c1

Using this account means I need to specify schema in the query, which is
what I've done in this commit.