Change code to meet rubocop rules.
A lot of tweaks here to get the rubocop task to run without complaint.

I'm changing some rubocop rules to meet the standard I think we should
follow.

- I'm not checking line length
- I'm allowing code like

```ruby
if params.valid?
 # do something
end
```

- I'm not requiring top-level comments
- I'm not checking quote type, either single or double is fine. I
  actually want to require double, but that would mean changing a bunch
  of require commands right now.

Of those, the if statement thing is probably the most contentious.
Rubocop really encourages single line conditionals, either through the
use of ternary or trailing conditionals. My preference is to just use a
single, consistent `if` statement.