Allow CachedCourseRepository to do multiple queries
Working towards a search that combines a subject and a le criteria, the
repo needs to be able to query on > 1 search parameter and return the
results that match all parameters. I'm doing this by getting each
result as a set and then intersecting them.

Implementation of this is gross, but working and covered by tests so
that it can be refactored later.

Doing this required converting SearchParameters from a Struct to a real
class because of the unexpected behavior that crops up when you convert
a Struct instance to an array.

Details on that here:
https://gist.github.com/IanWhitney/b5c62ba7bd8c975f4461

Or just look in the Ruby docs:
http://ruby-doc.org/core-2.1.5/Struct.html#method-i-to_a