Hide named finder methods behind 'where' method.
The goal of this branch is to get this app to the point where we can
query on 2 (or more) properties at a time. And, after playing around
with it a bit, these named finder methods on LiberalEducationCourse
seemed like they were going to get in the way of that.

If I left them as is, then we would eventually see this long chain of
class methods

```Ruby
LiberalEducationCourse.writing_intensive(true).diversified_core('AH').subject('HIST').all
```

Which might be how we want to interact with the class. I'm not sure. But
it certainly seemed simpler to do

```Ruby
LiberalEducationCourse.where(:writing_intensive: true, diversified_core:
'ah', subject: 'hist')
```

Or at least that seems better to me. It also avoids a couple problems I
ran into while trying to implement the method-chaining approach

- Class-level instance variables
- Having to tack `.all` at the end to actually return results (and wipe
  the class-level instance variable).

It's likely that there's a better way to implement the method-chaining
approach that I just missed.

A side-effect of implementing `where` is that the QueryParser can no
longer check to see if the search method is valid. That's fine, as that
part of QueryParser felt a little dumb anyway. LiberalEducationCourse
should be the one that determines what to return if given a weird
search.

The test of the bad endpoints mentioned in our readme have been moved to
LiberalEducationCourse, where that logic is now implemented.