Provide CourseSearch a collection of search_parameters
CourseSearch had been getting raw query strings and then parsing them.
Now it just gets a collection of search parameter objects that it can
pass on to its search target.

It does check to see if the search_params want to return all, in which
case it calls the target's all method. This could be better.