Add logic to query_parser for writing_intensive
This one feels a little weird, but I'm not sure where to put this logic
yet. The readme says that clients can search for writing_intensive
classes by including `writing_intensive=true`. But what we actually have
to search for are `writing_intensive=WI`. So I have to convert that
query string at some point. I used to do it in LiberalEducationCourse,
but not any more. So I'm doing it here.

Also, hiding the search_param attribute. It's no longer public