Move responsibility of finding search matches
Instead of having CachedCourseRepository match attributes and values,
just ask the SearchParameter if the course matches its criteria
(whatever those may be).

Why: In trying to refactor to a codebase that can easily accept this new
feature, I see the previous search implementation to be a serious
barrier. All it can do is see if the course's attribute matches the
parameter's attribute. This has led to weirdness already, as I have to
force the query_string of "writing_intensive=true" to a search for
for course with a writing_intensive value of "WI", instead of true.

If I followed that same course for "writing_intensive=false" then I'd
have to parse out the false and return some sort of search parameter
that looks for writing_intensive being nil.

And, even though I'm not implementing it yet, there's an idea on the
horizon of being able to search for a wider range of courses with a
query_string like "diversifed_core=all". That'd be even harder to parse
out into something the current search implementation would match.

By making CachedCourse just return things that `match?` I'm going to be able
to take advantage of polymorphism. I can pass in any object that
responds to `match?` and have it decide when to say true or false. This
is far more flexible.