Introduce FindAll type of SearchParameter
Why: Polymorphism!

As mentioned in the last commit, by moving the responsibility of
matching search results to the Search Parameter we would be able to use
any object that responds to `match?` when searching.

There was a lot of code that cared if the current request was a search
for 'all'. I was checking it in QueryParser (irrelevantly, actually), in
CourseSearch in SearchParameter and SearchParameters. And in every place
I had some weird if statement.

Replace that with FindAll which always returns `true` to `match?` and
all of those if statements go away.