Remove query_parser, add matcher_builder
Why: query_parser was kind of a weird remnant of an older approach. It
didn't really do anything except handle the odd case of
writing_intensive matching.

But with our matcher classes, that's no longer necessary. We just need a
class that gives us the an instance of the right matcher. Which is what
MatcherBuilder does. Note: all of these names are terrible and I intend
to fix them.