Implement writing_intensive=false
With the previous refactoring, the feature is now easy to implement.
I've added the MatchNoAttribute which matches targets whose value for
the attribute is nil. In other words, it matches courses that don't
have WI.

The if/elsif/else statment in matcher_builder is obviously terrible and
should go away soon. But I wanted to get the new feature test passing
before going after that bit of refactoring.