Extract conditional to polymorphism
Why: Polymorphism!

All that condintional logic in matcher_builder sucked and the last
couple of commits have been about trying to make it less terrible while
isolating each bit of conditional logic.

With that logic isolated, it's trivial to move the responsibility to the
Matchers themselves. They can decided, based on the search attributes, if they
should be built. This means MatcherBuilder just has to find a Matcher
that responds true to `build_me?`

This change required changes to the tests as the signature changed.
Also, it introduces a weird thing in `SearchParameters.matchers`.
`MatchAttributeValue` has to come after everything else because it will
match anything that has a search_type and search_param. This seems
brittle, but unlikely to cause problems right now. As we get more
requirements we might find a way to improve this design.