Remove vestigial CourseSearch class
This class used to do so much. So, so much. Then, as the code got
refactored, all of its responsibilities moved to their real homes. Like
kids heading off to start their adult lives. But classes can't hang out
and be empty nesters. Time for it to head off to the retirement village
in the sky.