Remove code that has been moved to ApiServiceSearching
After the last batch of refactoring, the code that handled searching was
totally separate from the code that retrieves or formats data. It has
been extracted out to a gem, so that we can use it on other api
services.

This commit updates this service to use the new gem.