Move cache configuration to initializer
As mentioned in commit 9df92e, removing jbuilder caused this cache
configuration code to fail. This is because we shouldn't be doing cache
configuration in the environment files. We should do it in an
initializer, which is called after ActiveSupport is loaded. So it'll run
whether or not jbuilder is installed