Show LE courses as of Fall 2015 term
This is a short term fix for issue #17 until we figure out how to solve
this feature long term. Instead of getting the most recent effective row
as of today, this query now gets the most recent row as of 9/8/2015.
Looking at the course catalog, that's the date used to for changes that
become effective with the Fall 2015 term.